+++ neo4j/logging.go
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2002-2019 "Neo4j,"
+ * Copyright (c) 2002-2020 "Neo4j,"
  * Neo4j Sweden AB [http://neo4j.com]
  *
  * This file is part of Neo4j.
@@ -20,11 +20,11 @@
 package neo4j
 
 import (
-	"fmt"
+	//"fmt"
 	"log"
 	"os"
-	"path/filepath"
-	"runtime"
+	//"path/filepath"
+	//"runtime"
 )
 
 // Logging is the interface that any provided logging target must satisfy for the neo4j
@@ -39,41 +39,6 @@ type Logging interface {
 	Warningf(message string, args ...interface{})
 	Infof(message string, args ...interface{})
 	Debugf(message string, args ...interface{})
-}
-
-func fileAndLineNumberOfCall() string {
-	_, file, line, ok := runtime.Caller(2)
-	if !ok {
-		file = "?"
-		line = 0
-	}
-	_, file = filepath.Split(file)
-
-	return fmt.Sprintf("%s:%d", file, line)
-}
-
-func errorf(logging Logging, message string, args ...interface{}) {
-	if logging != nil && logging.ErrorEnabled() {
-		logging.Errorf(fmt.Sprintf("%s: %s", fileAndLineNumberOfCall(), message), args...)
-	}
-}
-
-func warningf(logging Logging, message string, args ...interface{}) {
-	if logging != nil && logging.WarningEnabled() {
-		logging.Warningf(fmt.Sprintf("%s: %s", fileAndLineNumberOfCall(), message), args...)
-	}
-}
-
-func infof(logging Logging, message string, args ...interface{}) {
-	if logging != nil && logging.InfoEnabled() {
-		logging.Infof(fmt.Sprintf("%s: %s", fileAndLineNumberOfCall(), message), args...)
-	}
-}
-
-func debugf(logging Logging, message string, args ...interface{}) {
-	if logging != nil && logging.DebugEnabled() {
-		logging.Debugf(fmt.Sprintf("%s: %s", fileAndLineNumberOfCall(), message), args...)
-	}
 }
 
 // LogLevel is the type that default logging implementations use for available
