+++ neo4j/resultsummary.go
@@ -105,6 +105,74 @@ type ServerInfo interface {
 	Version() string
 }
 
+// Plan describes the actual plan that the database planner produced and used (or will use) to execute your statement.
+// This can be extremely helpful in understanding what a statement is doing, and how to optimize it. For more details,
+// see the Neo4j Manual. The plan for the statement is a tree of plans - each sub-tree containing zero or more child
+// plans. The statement starts with the root plan. Each sub-plan is of a specific operator, which describes what
+// that part of the plan does - for instance, perform an index lookup or filter results.
+// The Neo4j Manual contains a reference of the available operator types, and these may differ across Neo4j versions.
+type Plan interface {
+	// Operator returns the operation this plan is performing.
+	Operator() string
+	// Arguments returns the arguments for the operator used.
+	// Many operators have arguments defining their specific behavior. This map contains those arguments.
+	Arguments() map[string]interface{}
+	// Identifiers returns a list of identifiers used by this plan. Identifiers used by this part of the plan.
+	// These can be both identifiers introduced by you, or automatically generated.
+	Identifiers() []string
+	// Children returns zero or more child plans. A plan is a tree, where each child is another plan.
+	// The children are where this part of the plan gets its input records - unless this is an operator that
+	// introduces new records on its own.
+	Children() []Plan
+}
+
+// ProfiledPlan is the same as a regular Plan - except this plan has been executed, meaning it also
+// contains detailed information about how much work each step of the plan incurred on the database.
+type ProfiledPlan interface {
+	// Operator returns the operation this plan is performing.
+	Operator() string
+	// Arguments returns the arguments for the operator used.
+	// Many operators have arguments defining their specific behavior. This map contains those arguments.
+	Arguments() map[string]interface{}
+	// Identifiers returns a list of identifiers used by this plan. Identifiers used by this part of the plan.
+	// These can be both identifiers introduced by you, or automatically generated.
+	Identifiers() []string
+	// DbHits returns the number of times this part of the plan touched the underlying data stores/
+	DbHits() int64
+	// Records returns the number of records this part of the plan produced.
+	Records() int64
+	// Children returns zero or more child plans. A plan is a tree, where each child is another plan.
+	// The children are where this part of the plan gets its input records - unless this is an operator that
+	// introduces new records on its own.
+	Children() []ProfiledPlan
+}
+
+// Notification represents notifications generated when executing a statement.
+// A notification can be visualized in a client pinpointing problems or other information about the statement.
+type Notification interface {
+	// Code returns a notification code for the discovered issue of this notification.
+	Code() string
+	// Title returns a short summary of this notification.
+	Title() string
+	// Description returns a longer description of this notification.
+	Description() string
+	// Position returns the position in the statement where this notification points to.
+	// Not all notifications have a unique position to point to and in that case the position would be set to nil.
+	Position() InputPosition
+	// Severity returns the severity level of this notification.
+	Severity() string
+}
+
+// InputPosition contains information about a specific position in a statement
+type InputPosition interface {
+	// Offset returns the character offset referred to by this position; offset numbers start at 0.
+	Offset() int
+	// Line returns the line number referred to by this position; line numbers start at 1.
+	Line() int
+	// Column returns the column number referred to by this position; column numbers start at 1.
+	Column() int
+}
+
 type resultSummary struct {
 	sum    *db.Summary
 	cypher string
