--- neo4jx/summary_server_info.go
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) 2002-2019 "Neo4j,"
- * Neo4j Sweden AB [http://neo4j.com]
- *
- * This file is part of Neo4j.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *      http://www.apache.org/licenses/LICENSE-2.0
- *
- *  Unless required by applicable law or agreed to in writing, software
- *  distributed under the License is distributed on an "AS IS" BASIS,
- *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- *  See the License for the specific language governing permissions and
- *  limitations under the License.
- */
-
-package neo4j
-
-// ServerInfo contains basic information of the server.
-type ServerInfo interface {
-	// Address returns the address of the server.
-	Address() string
-	// Version returns the version of Neo4j running at the server.
-	Version() string
-}
-
-type neoServerInfo struct {
-	address string
-	version string
-}
-
-func (server *neoServerInfo) Address() string {
-	return server.address
-}
-
-func (server *neoServerInfo) Version() string {
-	return server.version
-}
